Split external
