+++ src/newClient/network.ts
@@ -1,22 +1,14 @@
 import { AxiosInstance } from 'axios';
 import { AssetResponse } from './types/asset';
-import {
-  CheckAddressRequest,
-  CheckAddressResponse,
-  NetworkAssetResponse,
-  NetworkChainResponse,
-  NetworkPriceResponse,
-  NetworkSnapshotResponse,
-  ExternalTransactionResponse,
-  DepositFilterRequest,
-  SnapshotFilterRequest,
-  NetworkInfo, ExchangeRateResponse
-} from './types/network';
+import { NetworkSnapshotRequest, NetworkAssetResponse, NetworkChainResponse, NetworkPriceResponse, NetworkSnapshotResponse, NetworkInfoResponse } from './types/network';
 import { buildClient } from './utils/client';
 
 // Public methods that need no permission
 // Detail: https://developers.mixin.one/docs/api/transfer/snapshots
 export const NetworkBaseClient = (axiosInstance: AxiosInstance) => ({
+  // Get network info
+  info: (): Promise<NetworkInfoResponse> => axiosInstance.get<unknown, NetworkInfoResponse>('/network'),
+
   // Get the list of all public chains supported by Mixin
   chains: (): Promise<NetworkChainResponse[]> => axiosInstance.get<unknown, NetworkChainResponse[]>('/network/chains'),
 
@@ -47,33 +39,21 @@ export const NetworkBaseClient = (axiosInstance: AxiosInstance) => ({
     return axiosInstance.get<unknown, NetworkPriceResponse>(`/network/ticker`, { params });
   },
 
+  // Get snapshot details by snapshot_id
+  // Make sure the dApp has already granted the SNAPSHOT:READ permission and set correct JWT in the request headers,
+  // to obtain the private fields like user_id, opponent_id, trace_id and data
+  snapshot: (snapshotID: string): Promise<NetworkSnapshotResponse> => axiosInstance.get<unknown, NetworkSnapshotResponse>(`/network/snapshots/${snapshotID}`),
+
   // Get snapshot records public information, which including transfers, deposits, withdrawals, etc
   // Make sure the dApp has already granted the SNAPSHOT:READ permission and set correct JWT in the request headers,
   // to obtain the private fields like user_id, opponent_id, trace_id and data
-  snapshots: (inputParams: SnapshotFilterRequest): Promise<NetworkSnapshotResponse[]> => {
+  snapshots: (inputParams: NetworkSnapshotRequest): Promise<NetworkSnapshotResponse[]> => {
     const params = {
       ...inputParams,
       order: inputParams.order || 'DESC'
     };
     return axiosInstance.get<unknown, NetworkSnapshotResponse[]>(`/network/snapshots`, { params });
   },
-
-  // Get snapshot details by snapshot_id
-  // Make sure the dApp has already granted the SNAPSHOT:READ permission and set correct JWT in the request headers,
-  // to obtain the private fields like user_id, opponent_id, trace_id and data
-  snapshot: (snapshotID: string): Promise<NetworkSnapshotResponse> => axiosInstance.get<unknown, NetworkSnapshotResponse>(`/network/snapshots/${snapshotID}`),
-
-  // Get public network-wide deposit records
-  deposit: (params: DepositFilterRequest): Promise<ExternalTransactionResponse[]> => axiosInstance.get<unknown, ExternalTransactionResponse[]>('/external/transactions', { params }),
-
-  // Check if an address belongs to Mixin
-  externalAddressesCheck: (params: CheckAddressRequest): Promise<CheckAddressResponse> => axiosInstance.get(`/external/addresses/check`, { params }),
-
-  // Get network info
-  info: (): Promise<NetworkInfo> => axiosInstance.get<unknown, NetworkInfo>('/network'),
-
-  // GET the list of all fiat exchange rates based on US Dollar
-  exchangeRates: (): Promise<ExchangeRateResponse> => axiosInstance.get<unknown, ExchangeRateResponse>('/external/fiats')
 });
 
 export const NetworkClient = buildClient(NetworkBaseClient);
