+++ src/newClient/types/network.ts
@@ -53,7 +53,7 @@ export interface NetworkPriceResponse {
   price_usd: number;
 }
 
-export interface NetworkInfo {
+export interface NetworkInfoResponse {
   assets: NetworkAssetResponse[];
   chains: NetworkChainResponse[];
   assets_count: string;
@@ -62,47 +62,9 @@ export interface NetworkInfo {
   type: string;
 }
 
-export interface ExternalTransactionResponse {
-  transaction_id: string;
-  created_at: string;
-  transaction_hash: string;
-  sender: string;
-  chain_id: string;
-  asset_id: string;
-  amount: string;
-  destination: string;
-  tag: string;
-  confirmations: string;
-  threshold: string;
-}
-
-export interface SnapshotFilterRequest {
+export interface NetworkSnapshotRequest {
   limit: number | string;
   offset: string;
   asset?: string;
   order?:string
-}
-
-export interface DepositFilterRequest {
-  limit: number | string;
-  offset: string;
-  asset?: string;
-  opponent?: string;
-  tag?: string;
-}
-
-export interface CheckAddressRequest {
-  asset: string;
-  destination: string;
-  tag?: string;
-}
-
-export interface CheckAddressResponse {
-  destination: string;
-  tag: string;
-}
-
-export interface ExchangeRateResponse {
-  code: string;
-  rate: string;
 }
